+++ metadeploy/api/models/cta.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+
+class ClickThroughAgreement(models.Model):
+    text = models.TextField()
